Added Gaia photometry search
Added cross matching with Gaia catalog. It's done "brute force" and I am sure there are smarter ways to do this.
The code create a photometry table with Gaia magnitude and errors for each CLAGN source.
Note that we also need to save somewhere the epoch of observations as well as the wavelengths of the filters.